Refactor:Projects screen align
